cranelift_assembler_x64

Struct Fixed

pub struct Fixed<R, const E: u8>(pub R);

A fixed register.

Some operands are implicit to the instruction and thus use a fixed register for execution. Because this assembler is generic over any register type (R), this wrapper provides a way to record the fixed register encoding we expect to use (E).

let fixed = Fixed::<u8, { gpr::enc::RAX }>(valid_reg);
assert_eq!(fixed.enc(), gpr::enc::RAX);

let fixed = Fixed::<u8, { gpr::enc::RAX }>(invalid_reg);
fixed.enc(); // Will panic because `invalid_reg` does not match `RAX`.

Tuple Fields
0: R
Implementations

impl<R, const E: u8> Fixed<R, E>

pub fn expected_enc(&self) -> u8

Return the fixed register encoding.

Regardless of what R is (e.g., pre-register allocation), we want to be able to know what this register should encode as.

pub fn to_string(&self, size: Option<Size>) -> String
where R: AsReg,

Return the register name at the given size.