cranelift_assembler_x64

Trait Registers

pub trait Registers {
    type ReadGpr: AsReg;
    type ReadWriteGpr: AsReg;
    type WriteGpr: AsReg;
    type ReadXmm: AsReg;
    type ReadWriteXmm: AsReg;
    type WriteXmm: AsReg;
}

A type set fixing the register types used in the assembler.

This assembler is parameterizable over register types; this allows the assembler users (e.g., Cranelift) to define their own register types independent of this crate.

Required Associated Types

type ReadGpr: AsReg

An x64 general purpose register that may be read.

type ReadWriteGpr: AsReg

An x64 general purpose register that may be read and written.

type WriteGpr: AsReg

An x64 general purpose register that may be written.

type ReadXmm: AsReg

An x64 SSE register that may be read.

type ReadWriteXmm: AsReg

An x64 SSE register that may be read and written.

type WriteXmm: AsReg

An x64 SSE register that may be written.